Update repository documentation for Markdown renderers consolidation
