--- cv_adapter/renderers/__init__.py
@@ -1,5 +1,0 @@
-from cv_adapter.renderers.base import BaseRenderer, RendererError
-from cv_adapter.renderers.markdown.markdown_renderer import MarkdownRenderer
-from cv_adapter.renderers.yaml_renderer import YAMLRenderer
-
-__all__ = ["BaseRenderer", "RendererError", "MarkdownRenderer", "YAMLRenderer"]
