--- cv_adapter/renderers/markdown/__init__.py
@@ -1,14 +1,0 @@
-# Markdown Renderers Package
-from .base_markdown_renderer import BaseMarkdownRenderer
-from .core_competences_renderer import CoreCompetencesRenderer
-from .markdown_list_renderer import MarkdownListRenderer
-from .markdown_renderer import MarkdownRenderer
-from .minimal_markdown_renderer import MinimalMarkdownRenderer
-
-__all__ = [
-    "BaseMarkdownRenderer",
-    "MarkdownListRenderer",
-    "MarkdownRenderer",
-    "MinimalMarkdownRenderer",
-    "CoreCompetencesRenderer",
-]
